zavrsene sve validacije

+++ centar-za-transfuziju-krvi/src/main/kotlin/com/isa/centarzatransfuzijukrvi/controller/AppointmentController.kt
@@ -2,20 +2,26 @@ package com.isa.centarzatransfuzijukrvi.controller
 
 import com.isa.centarzatransfuzijukrvi.model.Appointment
 import com.isa.centarzatransfuzijukrvi.model.Center
+import com.isa.centarzatransfuzijukrvi.model.DonorForm
 import com.isa.centarzatransfuzijukrvi.model.dto.*
 import com.isa.centarzatransfuzijukrvi.service.AppointmentService
+import com.isa.centarzatransfuzijukrvi.service.DonorFormService
 import com.isa.centarzatransfuzijukrvi.service.RegisteredUserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
+import java.time.Duration
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.ZoneId
 import java.util.*
 
 @RestController
 @RequestMapping(path = ["api"])
 @CrossOrigin(origins=["*"])
-class AppointmentController(@Autowired val appointmentService: AppointmentService, @Autowired val registeredUserService: RegisteredUserService) {
+class AppointmentController(@Autowired val appointmentService: AppointmentService, @Autowired val registeredUserService: RegisteredUserService, @Autowired val donorFormService: DonorFormService) {
 
     @PostMapping(path=["appointment/admin"], consumes = [MediaType.APPLICATION_JSON_VALUE], produces = [MediaType.APPLICATION_JSON_VALUE])
     fun createAdmin(@RequestBody unscheduledAppointment : AppointmentAdminDTO) : ResponseEntity<Appointment>{
@@ -37,8 +43,26 @@ class AppointmentController(@Autowired val appointmentService: AppointmentServic
     @PutMapping(path = ["/appointment"], consumes = [MediaType.APPLICATION_JSON_VALUE], produces = [MediaType.APPLICATION_JSON_VALUE])
     fun updateAppointment(@RequestBody dto: UpdateDTO) : ResponseEntity<Appointment> {
         var pacijent = registeredUserService.findByEmail(dto.email)
-        var povratna = appointmentService.updateAppointment(dto.id, pacijent)
-        return ResponseEntity(povratna, HttpStatus.OK)
+        var forme : ArrayList<DonorForm> = ArrayList()
+        var lista : ArrayList<Appointment> = ArrayList()
+
+        for(form in donorFormService.findAll()){
+            if(form.user == pacijent)
+                forme.add(form)
+        }
+        if(forme.size == 1) {
+            if (forme.get(0).question15 == "No") {
+                for (app in appointmentService.findAllRepo()) {
+                    if (app.donor == pacijent)
+                        lista.add(app)
+                }
+                if (lista.size == 0) {
+                    var povratna = appointmentService.updateAppointment(dto.id, pacijent)
+                    return ResponseEntity(povratna, HttpStatus.OK)
+                } else return ResponseEntity(null, HttpStatus.BAD_REQUEST)
+            } else return ResponseEntity(null, HttpStatus.UNAUTHORIZED)
+        }
+        else return ResponseEntity(null, HttpStatus.NOT_FOUND)
     }
 
     @GetMapping(path = ["/scheduledappointment/{email}/{role}/{enable}"],produces = [MediaType.APPLICATION_JSON_VALUE])
@@ -52,8 +76,25 @@ class AppointmentController(@Autowired val appointmentService: AppointmentServic
 
     @PutMapping(path = ["/deleteappointment"], consumes = [MediaType.APPLICATION_JSON_VALUE], produces = [MediaType.APPLICATION_JSON_VALUE])
     fun deleteAppointment(@RequestBody dto: UpdateDTO) : ResponseEntity<Appointment> {
-        var povratna = appointmentService.deleteAppointment(dto.id)
-        return ResponseEntity(povratna, HttpStatus.OK)
+        var lista : ArrayList<Appointment> = ArrayList()
+        var localDate = LocalDateTime.now()
+        var datum : Date = Date.from(localDate.atZone(ZoneId.systemDefault()).toInstant())
+        var pacijent = registeredUserService.findByEmail(dto.email)
+        for (app in appointmentService.findAllRepo()) {
+            if (app.donor == pacijent)
+                lista.add(app)
+        }
+        if(lista.size == 1) {
+            if((((lista.get(0).time.time - datum.time) / 1000) / 60) / 60 > 24){
+                println(lista.get(0).time)
+                println(datum)
+                println((((lista.get(0).time.time - datum.time) / 1000) / 60) / 60)
+                var povratna = appointmentService.deleteAppointment(dto.id)
+                return ResponseEntity(povratna, HttpStatus.OK)
+            }
+            else return ResponseEntity(null, HttpStatus.UNAUTHORIZED)
+        }
+        else return ResponseEntity(null, HttpStatus.OK)
     }
     @PostMapping(path=["appointment/user"], consumes = [MediaType.APPLICATION_JSON_VALUE], produces = [MediaType.APPLICATION_JSON_VALUE])
     fun queryAppointments(@RequestBody query : AppointmentSearchUserDTO) : ResponseEntity<List<AppointmentCenterUserDTO>>{
